backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.6
numpy>=1.24.0
openai-whisper>=20231117
chatterbox-tts>=0.1.2
